requirements.txt
numpy~=1.24.2
matplotlib~=3.7.0
scikit-learn~=1.2.1
opencv-python~=4.7.0.72
mediapipe~=0.9.1.0
seaborn~=0.12.2
tensorflow-gpu~=2.10.0
transformers~=4.26.1
keras~=2.10.0